Add proper server check for localhost use cases
The Prefix option was removed, because that was a temporary
workaround for getting the example OpenAPI spec to work with
the validator. The example spec now includes https://localhost:9443
as the example host, which is the same as in the config.json.
Furthermore, we're now properly reconstructing the URL to
use when determining the route from the OpenAPI spec.